requirements.txt
Django==1.10.5
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.45.0
celery==3.1.25
celery-with-redis==3.0
chardet==2.3.0
diff-match-patch==20121119
dj-static==0.0.6
django-autocomplete-light==3.2.1
django-celery==3.2.1
django-csv-exports==1.0.3
django-debug-toolbar==1.6
django-grappelli==2.9.1
django-redis==4.7.0
django-redis-cache==1.7.1
django-suit==0.2.24
djangorestframework==3.5.4
gunicorn==19.6.0
heroku==0.1.4
kombu==3.0.37
psycopg2==2.6.2
python-dateutil==1.5
pytz==2016.10
redis==2.10.5
requests==2.13.0
sqlparse==0.2.2
static3==0.7.0
whitenoise==3.3.0
